requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
